+++ R/emission.R
@@ -4,15 +4,15 @@
 #'
 #' @format matrix of emission
 #'
+#' @param inventory a inventory raster from read
+#' @param grid grid information
+#' @param mm pollutant molar mass
+#' @param aerosol TRUE for aerosols and FALSE (defoult) for gazes
 #' @param total list of total emission
 #' @param pol pollutant name
 #' @param area list of area sources or matrix with a spatial distribution
-#' @param grid grid information
-#' @param inventory a inventory raster from read
-#' @param mm pollutant molar mass
-#' @param aerosol TRUE for aerosols and FALSE (defoult) for gazes
 #' @param plot TRUE for plot the final emissions
-#' @param positive TRUE (defoult) to check negative values and replace for zero
+#' @param check TRUE (defoult) to check negative and NA values and replace it for zero
 #' @param verbose display additional information
 #'
 #' @return a vector of emissions in MOL / mk2 h for gases and ug / m2 s for aerosols.
@@ -45,11 +45,16 @@
 #' SP     <- areaSource(shape[22,1],raster,grid,name = "SP")
 #' RJ     <- areaSource(shape[17,1],raster,grid,name = "RJ")
 #'
-#' e_CO   <- emission(TOTAL,"CO",list(SP = SP, RJ = RJ),grid,mm=28)
+#' e_CO   <- emission(total = TOTAL,
+#'                    pol = "CO",
+#'                    area = list(SP = SP, RJ = RJ),
+#'                    grid = grid,
+#'                    mm = 28)
 #'
 
-emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = FALSE,
-                     plot = FALSE, positive = TRUE,verbose = TRUE){
+emission <- function(inventory = NULL,grid,mm = 1, aerosol = FALSE,check = TRUE,
+                     total, pol, area,
+                     plot = FALSE,verbose = TRUE){
 
   if(!is.null(inventory)){
     if(verbose){
@@ -75,10 +80,14 @@ emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = FAL
     else{
       ##  mol km-2 h-1
       VAR_e   =  units::set_units(VAR_e,"g km-2 h-1")
-      suppressWarnings( units::install_symbolic_unit("MOL") )
-      MOL <- units::as_units("MOL")
-      conversao <- units::as_units(1/mm, "MOL g-1")
-      VAR_e     <- VAR_e * conversao
+      ### removing deprecated install_symbolic_unit function for units v0.8-0+
+      # suppressWarnings( units::install_symbolic_unit("MOL") )
+      # MOL <- units::as_units("MOL")
+      # conversao <- units::as_units(1/mm, "MOL g-1")
+      # VAR_e     <- VAR_e * conversao
+      units::remove_unit("MOL")
+      units::install_unit("MOL", paste(mm,"g"))
+      VAR_e <- units::set_units(VAR_e,"MOL km-2 h-1")
     }
 
     if(plot == TRUE){
@@ -108,7 +117,7 @@ emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = FAL
 
       print(a)
     }
-    if(positive){
+    if(check){
       VAR_e <- check_positive(VAR_e,pol)
       return(VAR_e)
     }else{
@@ -151,7 +160,9 @@ emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = FAL
     }
     VAR_e[is.na(VAR_e)]     <- 0
 
-    VAR_e <- rasterSource(VAR_e,grid,verbose = FALSE)
+    # old code
+    # VAR_e <- rasterSource(VAR_e,grid,verbose = FALSE)
+    VAR_e <- raster_to_ncdf(VAR_e)
 
     # put the units (to back the unit)
     VAR_e <- VAR_e * unidade
@@ -168,11 +179,14 @@ emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = FAL
   }
   else{
     #  mol km^-2 hr^-1
-    if(exists("ud_units$MOL"))
-      units::remove_symbolic_unit("MOL") # nocov
-    if(!exists("ud_units$MOL"))
-      # units::install_conversion_constant("MOL", "g", const = mm) # old conversion function
-      units::install_unit("MOL", paste(mm,"g"))                    # new conversion function
+    ### removing deprecated install_symbolic_unit function for units v0.8-0+
+    ### if(exists("ud_units$MOL"))
+    ###   units::remove_symbolic_unit("MOL") # nocov
+    ### if(!exists("ud_units$MOL"))
+    ###   # units::install_conversion_constant("MOL", "g", const = mm) # old conversion function
+    ###   units::install_unit("MOL", paste(mm,"g"))                    # new conversion function
+    units::remove_unit("MOL")
+    units::install_unit("MOL", paste(mm,"g"))
     VAR_e   =  units::set_units(VAR_e,"MOL/d")
     VAR_e   =  units::set_units(VAR_e,"MOL/h")
     VAR_e   =  VAR_e / dx^2
@@ -199,7 +213,7 @@ emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = FAL
 
     print(a)
   }
-  if(positive){
+  if(check){
     VAR_e <- check_positive(VAR_e,pol)
     return(VAR_e)
   }else{
@@ -208,19 +222,35 @@ emission <- function(total,pol,area,grid, inventory = NULL,mm = 1, aerosol = FAL
 }
 
 check_positive <- function(emiss,pol = '?'){
-  warn <- FALSE
+  has_NAs  <- FALSE
+  negative <- FALSE
+  n_na  = 0
+  n_neg = 0
   for(i in 1:length(emiss)){
     if(is.na(drop_units(emiss[i]))){
-      warn <- TRUE                         # nocov
+      has_NAs  <- TRUE                     # nocov
       emiss[i] = 0                         # nocov
+      n_na     = n_na + 1                  # nocov
     }else{
       if(drop_units(emiss[i]) < 0){
-        warn <- TRUE                       # nocov
+        negative <- TRUE                   # nocov
         emiss[i] = 0                       # nocov
+        n_neg    = n_neg + 1               # nocov
       }
     }
   }
-  if(warn)
-    warning('Negative or NA values found, replaced by zero in ',pol) # nocov
+  if(negative)
+    warning(n_neg,' Negative values found, replaced by zero in ',pol) # nocov
+  if(has_NAs)
+    warning(n_na,' NA values found, replaced by zero in ',pol)        # nocov
   return(emiss)
+}
+
+raster_to_ncdf <- function(r,na_value = 0){
+  N_times <- dim(r)[3]
+  a       <- array(na_value,c(dim(r)[2],dim(r)[1],N_times))
+  for(i in 1:N_times){
+    a[,,i] <- as.matrix(t(raster::flip(r[[i]],2)))
+  }
+  return(a)
 }