+++ R/rasterSource.R
@@ -16,8 +16,7 @@
 #'
 #' @export
 #'
-#' @import raster sp
-#' @importFrom rgdal project
+#' @import raster sp sf
 #'
 #' @examples
 #' grid  <- gridInfo(paste(system.file("extdata", package = "EmissV"),"/wrfinput_d01",sep=""))
@@ -36,10 +35,13 @@ rasterSource <- function(r,grid,nlevels="all",conservative = TRUE,verbose = TRUE
     dy    <- grid$DX*1000            # using meters
     ncols <- grid$Horizontal[1]
     nrows <- grid$Horizontal[2]
-    projcoords <- rgdal::project(grid$coords,
-                                 grid$geogrd.proj)
-    # projcoords <- sp::spTransform(x      = grid$coords,
-    #                               CRSobj = grid$geogrd.proj)
+
+    # projcoords <- rgdal::project(grid$coords,grid$geogrd.proj)
+    pontos     <- sf::st_multipoint(x = grid$coords, dim = "XY")
+    coords     <- sf::st_sfc(x = pontos, crs = "+proj=longlat")
+    transform  <- sf::st_transform(x = coords, crs = grid$geogrd.proj)
+    projcoords <- sf::st_coordinates(transform)[,1:2]
+
     xmn <- projcoords[1,1] - dx/2.0  # Left border
     ymx <- projcoords[1,2] + dy/2.0  # upper border
     xmx <- xmn + ncols*dx            # Right border
@@ -70,7 +72,7 @@ rasterSource <- function(r,grid,nlevels="all",conservative = TRUE,verbose = TRUE
     box_ll <- suppressWarnings( projectRaster(box, crs='+proj=longlat') )
     r      <- raster::crop(r,raster::extent(box_ll))
     if(conservative)
-      total_box <- cellStats(r,"sum",na.rm=TRUE)
+      total_box <- cellStats(r,"sum",na.rm=TRUE) # nocov
 
     r    <- suppressWarnings(raster::projectRaster(r,crs = raster::crs(box))) # to the new projection
     X    <- raster::resample(r,box,method = "bilinear")                       # non-conservative transformation
@@ -78,9 +80,9 @@ rasterSource <- function(r,grid,nlevels="all",conservative = TRUE,verbose = TRUE
     X    <- raster::t(X)
     X    <- raster::as.matrix(X)
     X[is.na(X)] <- 0             # for low resolution input data
-
+    # to conserve mass
     if(conservative)
-      X    <- X * total_box/sum(X) # to conserve mass
+      X    <- X * total_box/sum(X) # nocov
 
     if(verbose)
       cat(paste("Grid output:",grid$Horizontal[1],